autoGenVideo/requirements.txt
# AI / LLM
anthropic>=0.25.0
openai>=1.30.0
elevenlabs>=1.0.0

# Web Framework
fastapi>=0.111.0
uvicorn>=0.30.0

# Scheduling
apscheduler>=3.10.4

# Database
sqlalchemy>=2.0.30
python-dotenv>=1.0.1

# Video
moviepy>=1.0.3
Pillow>=10.3.0
ffmpeg-python>=0.2.0
pydub>=0.25.1
gTTS>=2.5.1

# Scraping / Data
requests>=2.31.0
beautifulsoup4>=4.12.3
pytrends>=4.9.2
praw>=7.7.1
feedparser>=6.0.11
google-api-python-client>=2.130.0
google-auth-oauthlib>=1.2.0

# Platform
instagrapi>=2.0.0
slack-sdk>=3.27.2
tenacity>=8.3.0
